# Conflicts:
#	CHANGELOG.md
#	package-lock.json
#	package.json
#	src/app/bots/bots-list/bots-list.component.ts
#	src/app/bots/bots-list/clone-bot/clone-bot.component.ts
#	src/app/canned-responses/canned-responses-list.component.html
#	src/app/canned-responses/canned-responses-list.component.ts
#	src/app/components/sidebar-user-details/sidebar-user-details.component.scss
#	src/app/departments/departments.component.html
#	src/app/groups/groups.component.html
#	src/app/integrations/integrations.component.html
#	src/app/integrations/integrations.component.ts
#	src/app/integrations/integrations.module.ts
#	src/app/integrations/utils.ts
#	src/app/knowledge-bases/modals/modal-preview-knowledge-base/modal-preview-knowledge-base.component.ts
#	src/app/knowledge-bases/modals/modal-upload-file/modal-upload-file.component.html
#	src/app/services/faq-kb.service.ts
#	src/app/widget_components/widget-set-up/widget-set-up.component.html
#	src/app/ws_requests/ws-requests-msgs/ws-requests-msgs.component.html
#	src/app/ws_requests/ws-requests-msgs/ws-requests-msgs.component.scss
#	src/app/ws_requests/ws-requests-msgs/ws-requests-msgs.component.ts
#	src/assets/i18n/ar.json
#	src/assets/i18n/az.json
#	src/assets/i18n/de.json
#	src/assets/i18n/en.json
#	src/assets/i18n/es.json
#	src/assets/i18n/fr.json
#	src/assets/i18n/it.json
#	src/assets/i18n/kk.json
#	src/assets/i18n/pt.json
#	src/assets/i18n/ru.json
#	src/assets/i18n/sr.json
#	src/assets/i18n/sv.json
#	src/assets/i18n/tr.json
#	src/assets/i18n/uk.json
#	src/assets/i18n/uz.json

+++ src/app/canned-responses/canned-responses-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, HostListener, OnInit } from '@angular/core';
+import { Component, HostListener, OnDestroy, OnInit } from '@angular/core';
 import { CannedResponsesService } from '../services/canned-responses.service';
 import { TranslateService } from '@ngx-translate/core';
 import { NotifyService } from '../core/notify.service';
@@ -10,12 +10,16 @@ import { AppConfigService } from '../services/app-config.service';
 import { RoleService } from 'app/services/role.service';
 import { BrandService } from 'app/services/brand.service';
 import { URL_canned_responses_doc } from 'app/utils/util';
+import { Subject } from 'rxjs';
+import { RolesService } from 'app/services/roles.service';
+import { takeUntil } from 'rxjs/operators';
+import { PERMISSIONS } from 'app/utils/permissions.constants';
 @Component({
   selector: 'appdashboard-canned-responses-list',
   templateUrl: './canned-responses-list.component.html',
   styleUrls: ['./canned-responses-list.component.scss']
 })
-export class CannedResponsesListComponent implements OnInit {
+export class CannedResponsesListComponent implements OnInit, OnDestroy {
 
   public canned_responses_docs_url = URL_canned_responses_doc
   displayModal_AddEditResponse = 'none';
@@ -35,8 +39,15 @@ export class CannedResponsesListComponent implements OnInit {
   baseUrl: string;
   isChromeVerGreaterThan100: boolean
   public hideHelpLink: boolean;
-
   currentUserId: string;
+  private unsubscribe$: Subject<any> = new Subject<any>();
+  isAuthorized = false;
+  permissionChecked = false;
+
+  PERMISSION_TO_UPDATE: boolean;
+  PERMISSION_TO_CREATE: boolean;
+  PERMISSION_TO_DELETE: boolean;
+
 
   constructor(
     public cannedResponsesService: CannedResponsesService,
@@ -49,6 +60,7 @@ export class CannedResponsesListComponent implements OnInit {
     public appConfigService: AppConfigService,
     public roleService: RoleService,
     public brandService: BrandService,
+    public rolesService: RolesService
   ) {
     const brand = brandService.getBrand(); 
     this.hideHelpLink= brand['DOCS'];
@@ -56,16 +68,24 @@ export class CannedResponsesListComponent implements OnInit {
 
   ngOnInit() {
     // this.auth.checkRoleForCurrentProject();
-    this.roleService.checkRoleForCurrentProject('canned-list')
+    // this.roleService.checkRoleForCurrentProject('canned-list')
     this.getResponses();
     this.translateNotificationMsgs();
     // this.getMainPanelAndSetOverflow();
     this.listenSidebarIsOpened();
     this.getImageStorage();
     this.getBrowserVersion();
+    this.checkPermissions();
+    this.listenToProjectUser()
     this.getCurrentUser()
   }
 
+
+  ngOnDestroy() {
+    this.unsubscribe$.next();
+    this.unsubscribe$.complete();
+  }
+  
    getCurrentUser() {
     this.auth.user_bs.subscribe((user) => {
       console.log('[CANNED-RES-LIST] - LoggedUser ', user);
@@ -75,6 +95,73 @@ export class CannedResponsesListComponent implements OnInit {
       }
     });
   }
+
+  async checkPermissions() {
+    const result = await this.roleService.checkRoleForCurrentProject('canned-responses')
+    console.log('[CANNED-RES-LIST] result ', result)
+    this.isAuthorized = result === true;
+    this.permissionChecked = true;
+    console.log('[CANNED-RES-LIST] isAuthorized ', this.isAuthorized)
+    console.log('[CANNED-RES-LIST] permissionChecked ', this.permissionChecked)
+  }
+
+    listenToProjectUser() {
+      this.rolesService.listenToProjectUserPermissions(this.unsubscribe$);
+      this.rolesService.getUpdateRequestPermission()
+        .pipe(takeUntil(this.unsubscribe$))
+        .subscribe(status => {
+          console.log('[CANNED-RES-LIST] - listenToProjectUser-  Role:', status.role);
+          console.log('[CANNED-RES-LIST] - listenToProjectUser - Permissions:', status.matchedPermissions);
+
+          // PERMISSION TO UPDATE
+          if (status.role !== 'owner' && status.role !== 'admin' && status.role !== 'agent') {
+            if (status.matchedPermissions.includes(PERMISSIONS.CANNED_RESPONSES_UPDATE)) {
+  
+              this.PERMISSION_TO_UPDATE = true
+              console.log('[CANNED-RES-LIST] - listenToProjectUser PERMISSION_TO_UPDATE ', this.PERMISSION_TO_UPDATE);
+            } else {
+              this.PERMISSION_TO_UPDATE = false
+              console.log('[CANNED-RES-LIST] - listenToProjectUser PERMISSION_TO_UPDATE ', this.PERMISSION_TO_UPDATE);
+            }
+          } else {
+            this.PERMISSION_TO_UPDATE = true
+            console.log('[CANNED-RES-LIST] - listenToProjectUser Project user has a default role ', status.role, 'PERMISSION_TO_UPDATE ', this.PERMISSION_TO_UPDATE);
+          }
+
+          // PERMISSION TO CREATE
+          if (status.role !== 'owner' && status.role !== 'admin' && status.role !== 'agent') {
+            if (status.matchedPermissions.includes(PERMISSIONS.CANNED_RESPONSES_CREATE)) {
+  
+              this.PERMISSION_TO_CREATE = true
+              console.log('[CANNED-RES-LIST] - listenToProjectUser PERMISSION_TO_CREATE 1', this.PERMISSION_TO_CREATE);
+            } else {
+              this.PERMISSION_TO_CREATE = false
+              console.log('[CANNED-RES-LIST] - listenToProjectUser PERMISSION_TO_CREATE 2', this.PERMISSION_TO_CREATE);
+            }
+          } else {
+            this.PERMISSION_TO_CREATE = true
+            console.log('[CANNED-RES-LIST] - listenToProjectUser Project user has a default role ', status.role, 'PERMISSION_TO_CREATE ', this.PERMISSION_TO_CREATE);
+          }
+
+           // PERMISSION TO 
+          if (status.role !== 'owner' && status.role !== 'admin' && status.role !== 'agent') {
+            if (status.matchedPermissions.includes(PERMISSIONS.CANNED_RESPONSES_DELETE)) {
+  
+              this.PERMISSION_TO_DELETE = true
+              console.log('[CANNED-RES-LIST] - listenToProjectUser PERMISSION_TO_DELETE 1', this.PERMISSION_TO_DELETE);
+            } else {
+              this.PERMISSION_TO_DELETE = false
+              console.log('[CANNED-RES-LIST] - listenToProjectUser PERMISSION_TO_DELETE 2', this.PERMISSION_TO_DELETE);
+            }
+          } else {
+            this.PERMISSION_TO_DELETE = true
+            console.log('[CANNED-RES-LIST] - listenToProjectUser Project user has a default role ', status.role, 'PERMISSION_TO_DELETE ', this.PERMISSION_TO_DELETE);
+          }
+
+  
+          // You can also check status.role === 'owner' if needed
+        });
+    }
 
   getBrowserVersion() {
     this.auth.isChromeVerGreaterThan100.subscribe((isChromeVerGreaterThan100: boolean) => {
@@ -148,7 +235,7 @@ export class CannedResponsesListComponent implements OnInit {
   getResponses() {
     // this.contactsService.getLeads(this.queryString, this.pageNo).subscribe((leads_object: any) => {
     this.cannedResponsesService.getCannedResponses().subscribe((responses: any) => {
-      // console.log('[CANNED-RES-LIST] - GET CANNED RESP - RES ', responses);
+     console.log('[CANNED-RES-LIST] - GET CANNED RESP - RES ', responses);
       if (responses) {
         this.responsesList = responses;
 
@@ -192,6 +279,10 @@ export class CannedResponsesListComponent implements OnInit {
 
 
   openConfirmDialogDeleteResponse(cannedresponseid) {
+    if (!this.PERMISSION_TO_DELETE) { 
+      this.notify.presentDialogNoPermissionToPermomfAction()
+      return
+    }
     this.displayModalDeleteResponse = 'block'
     this.cannedResponseIdToDelete = cannedresponseid;
   }
@@ -201,46 +292,63 @@ export class CannedResponsesListComponent implements OnInit {
   }
 
   deleteCannedResponse() {
-    this.cannedResponsesService.deleteCannedResponse(this.cannedResponseIdToDelete).subscribe((responses: any) => {
-      this.logger.log('[CANNED-RES-LIST] - DELETE CANNED RESP - RES ', responses);
+    if (this.PERMISSION_TO_DELETE) {
+      this.cannedResponsesService.deleteCannedResponse(this.cannedResponseIdToDelete).subscribe((responses: any) => {
+        this.logger.log('[CANNED-RES-LIST] - DELETE CANNED RESP - RES ', responses);
 
-    }, (error) => {
-      this.logger.error('[CANNED-RES-LIST] - DELETE CANNED RESP - ERROR  ', error);
-      this.displayModalDeleteResponse = 'none'
-      this.notify.showWidgetStyleUpdateNotification(this.deleteErrorMsg, 4, 'report_problem');
-    }, () => {
-      this.logger.log('[CANNED-RES-LIST] - DELETE CANNED RESP * COMPLETE *');
-      this.notify.showWidgetStyleUpdateNotification(this.deleteSuccessMsg, 2, 'done');
-      // this.getResponses()
-      this.displayModalDeleteResponse = 'none'
-      for (var i = 0; i < this.responsesList.length; i++) {
-        if (this.responsesList[i]._id === this.cannedResponseIdToDelete) {
-          this.responsesList.splice(i, 1);
-          i--;
+      }, (error) => {
+        this.logger.error('[CANNED-RES-LIST] - DELETE CANNED RESP - ERROR  ', error);
+        this.displayModalDeleteResponse = 'none'
+        this.notify.showWidgetStyleUpdateNotification(this.deleteErrorMsg, 4, 'report_problem');
+      }, () => {
+        this.logger.log('[CANNED-RES-LIST] - DELETE CANNED RESP * COMPLETE *');
+        this.notify.showWidgetStyleUpdateNotification(this.deleteSuccessMsg, 2, 'done');
+        // this.getResponses()
+        // this.cannedResponseIdToDelete = null
+        this.displayModalDeleteResponse = 'none'
+        for (var i = 0; i < this.responsesList.length; i++) {
+          if (this.responsesList[i]._id === this.cannedResponseIdToDelete) {
+            this.responsesList.splice(i, 1);
+            i--;
+          }
         }
-      }
+        
+      });
 
-    });
+    } else {
+      this.notify.presentDialogNoPermissionToPermomfAction()
+    }
   }
 
 
   presentResponseModal_inAddMode() {
-    this.selectCannedResponseId = null;
-    this.displayModal_AddEditResponse = 'block';
-    this.modalMode = 'add';
-    this.logger.log('[CANNED-RES-LIST] - displayModal ', this.displayModal_AddEditResponse, ' in Mode', this.modalMode);
+    if (this.PERMISSION_TO_CREATE) {
+      this.selectCannedResponseId = null;
+      this.displayModal_AddEditResponse = 'block';
+      this.modalMode = 'add';
+      this.logger.log('[CANNED-RES-LIST] - displayModal ', this.displayModal_AddEditResponse, ' in Mode', this.modalMode);
+    } else {
+      this.notify.presentDialogNoPermissionToPermomfAction()
+    }
   }
 
-  presentResponseModal_inEditMode(cannedresponseid: string, createdby) {
+  presentResponseModal_inEditMode(cannedresponseid: string, createdby: string) {
+    if (!this.PERMISSION_TO_UPDATE) {
+      this.notify.presentDialogNoPermissionToPermomfAction()
+      return
+    }
     if (createdby !== this.currentUserId) {
       return
     }
+
+    console.log('[CANNED-RES-LIST] - createdby ', createdby, ' currentUserId ' , this.currentUserId);
+    
     this.getScrollPos();
     this.selectCannedResponseId = cannedresponseid;
     this.displayModal_AddEditResponse = 'block';
     this.modalMode = 'edit';
-    this.logger.log('[CANNED-RES-LIST] - displayModal ', this.displayModal_AddEditResponse, ' in Mode', this.modalMode, ' canned-response-id', cannedresponseid);
-
+    console.log('[CANNED-RES-LIST] - displayModal ', this.displayModal_AddEditResponse, ' in Mode', this.modalMode, ' canned-response-id', cannedresponseid);
+    
   }
 
   getScrollPos() {
